+++ src/starbase/__init__.py
@@ -1,6 +1,6 @@
 __title__ = 'starbase.__init__'
-__version__ = '0.3'
-__build__ = 0x000003
+__version__ = '0.3.1'
+__build__ = 0x00000b
 __author__ = 'Artur Barseghyan'
 __copyright__ = 'Copyright (c) 2013 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'